Need to get the female worms by age groups output to match those of the stochastic R model

+++ CHost.h
@@ -17,8 +17,8 @@ public:
 
 	// Members.
 	double birthDate, deathDate;
-	double femaleWorms, totalWorms;
-	double freeliving;
+	int femaleWorms, totalWorms;
+	int freeliving;
 };
 
 #endif /* CHOST_H_ */
